contract_service/requirements.txt
fastapi
uvicorn
kubernetes
pydantic
psycopg2-binary
